Remove broken SCSS rule
When working on the new CSS compiler, we found this broken SCSS rule,
which simplified looks like this:
```scss
.class::before {
  &.other-class {
    top: 30px;
  }
}
```

It currently compiles to `.class.other-class::before`. In the new SASS
compiler, it compiles to `.class::before.other-class` which is invalid.
It's really confusing why this works _currently_, and it's the only
of this in our code base.

Luckily for us, the rule seems obsolete (I suspect that something in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/135038 might have
fixed it).