workhorse: Downgrade incomplete multipart uploads to 400 errors
In https://gitlab.com/gitlab-com/gl-infra/production/-/issues/17587,
our 500 error rates rose when incomplete multipart uploads were sent,
but this is really a client error, not a server error. This commit
downgrades that to a 400 error by doing an ugly string comparison in
`multipart.NextPart` to check for this condition.

Changelog: changed