Reuse existing projects & groups resolvers for Organization
Here we have changed the base class of Organizations::GroupsResolver and
Organizations::ProjectsResolver from BaseResolver to GroupsResolver and
ProjectsResolver respectively to use the existing functionalities of
both the resolvers.

We also need a `search` argument in the Organizations::ProjectsResolver
and changing its base class added this argument as well.

The `GroupsResolver` was missing the sort argument which has been added
as well.

Changelog: added