WorkItems rolledup dates: make fixed dates nullable on graphql
When working on the frontend, it was noticed that there was no way to
empty the fixed dates (set to null). We decided to change the logic in
the backend to be based on the `start_date_is_fixed` only. In other
words, when `start_date_is_fixed` is true, we assume the frontend is
sending the desired fixed dates in `start_date_fixed` field to be set on
the backend. (samething for `due_date`)

Changelog: added
EE: true