Add Deployment Approvals table to Details page
We get to reuse the mutliple approval rules table built out for the
environments page, but the rest of the form needs to be re-done a bit.
We do get to use GraphQL mutations for the actions, though, and I think
eventually we are funnelling users here so those other components will
be deleted eventually.

DOES NOT SUPPORT DISPLAY UNIFIED APPROVAL RULES. I'm not sure if we need
to, as we're deprecating the functionality in 17.0, and the main
environments page still handles it.