Add metrics for keep-around ref creation
Keep-around refs are used to stop dangling commits from being garbage
collected so that various UI elements on gitlab continue to work. The
problem is that it is impossible to evaluate where these refs are
created after the fact, all you know is that they got created. So here
we add metrics that can be explicitly bucketed by the source of the
keep-around.

Changelog: added