Fix NoMethodError caused by nil source attribute
Only GitLab generated SBOMs can be assumed to have a non-nil source attribute.
If we assume that they're always non-nil, we run into situations where we
access an attribute like the operating system name on a nil object. This commit
removes this risk by delegating the attributes to the source, and allows for a
nil value to be safely handled.