Add a link to a meta feedback issue for gitlab-housekeeper
As suggested in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/144559#note_1772656778
it would be helpful for people to have a link to somewhere they can
provide feedback for housekeeper MRs. In future we might consider
automating the creation of a feedback for every keep. For now I don't
want to add extra friction for creating a keep so I've just created a
single meta issue. We'll see how we go with that before we decide if we
need more issues.