Fix boolean and number input types
Boolean and number input types were not be respected by CI
interpolation. They were being cast to strings when interpolated by
`gsub`. This commit fixes that by replacing the entire node with the
input value when the interpolation block is the entire length of the
node. In other cases, we can assume that the input is being inserted
into a string, so it's safe for us to use `gsub` to cast the value to a
string and perform the insertion.

Changelog: fixed
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/434826