Purge old `security_scans` records on self-hosted installations
This change introduces a background migration to purge `security_scans`
records associated with `security_findings` without `finding_data` set.

These kinds of records do not exist on GitLab.com as the `finding_data`
column was introduced in '15.6' and the records created before that
milestone are already removed from the database by the data retention
policy of 3 months on the `security_findings` table but there is no
guarantee for the self-hosted installations.

Therefore, the background migration introduced in this MR will only run
for self-hosted installations.

Changelog: added