Add `.actor_from_id` to check feature flags with model id
Internally, flipper checks the model class name and model id in order to
determine if a feature is enabled for an actor. If we already have an
actor's ID, there is no need to make a database query for that actor for
the sole purpose of performing a feature flag check. Here we add the
`actor_from_id` method to models, which returns a wrapper around the
class name and model ID that builds a `flipper_id`. This lets us perform
one query for a feature flag check instead of two. The only downside of
this approach is that we might perform feature flag checks for actors
which no longer exist. I think this downside is negligible because:

1. If an administrator explicitly enabled the flag for a given actor,
   then people must be using that object and it's very unlikely for it
   to be deleted.
2. In the contexts where we have an ID and only want to perform a
   feature flag check, we are probably passing that ID downstream to
   another process (like a sidekiq worker) and that process will perform
   an existence check on the object.