Fix Duo Chat messages being saved twice
We support two response handling workflows. The default response handler
is always executed, and then there's an optional streaming response
handler. Both of them currently save the Chat message, which results in
duplicated messages in our Chat storage. To address this, we introduce a
new toggle, and only save the message on the default response handler,
and skip it on the streaming one