Add bbm dictionary content to cop cache key
The BackgroundMigration/DictionaryFile relies upon the contents
of the files in db/docs/batched_background_migrations/ as well as
the Ruby source file of the background migration.

Currently, the Rucocop results cache will not be updated if the
contents of the docs folder changes, which leads to false positives
and negatives.

This change includes the contents of the docs folder in the cache
checksum, so any changes in that directory will bust the result cache.