CurrentUserMode with session, load known attributes within ActiveSession
The original MR(Show admin mode within active sessions view) has
been reverted due to failures:
https://gitlab.com/gitlab-org/release/tasks/-/issues/8332#note_1748203585 on staging-canary.
See discussion https://gitlab.com/gitlab-org/gitlab/-/merge_requests/142158#note_1748221369
It was agreed to implement it in two steps, and this is step one.

Within this MR CurrentUserMode is extend with session parameter and
within load_raw_session only known attributes of ActiveSession
are loaded. This fixes the root cause for the crash within
staging-canary.
In the future new attributes can be added without the risk of a crash
as only known attributes are passed to the constructor.