Cache list of conflicts when serializing diffs
When we serialize diffs and the MR has conflicts, we make a
`ListConflictFiles` RPC call to Gitaly. We use this to determine
what type of conflict file/s in the diffs is encountered.

This RPC can take a while if
we're requesting this every request and there are no actual
changes.

To improve performance for succeeding requests, we introduce a new
`#conflicts_with_types` helper which will cache the result and
will use the cached result until source/target branch changes.

Caching functionality is behind `cached_conflicts_with_types`
feature flag.