Refactor BranchRules::BaseService to handle common pattern
In the EE version there are custom branch rules which need to be treated
differently. The branch rule service classes have a common pattern which
detects which type of branch rule is currently being passed.

This change extracts that logic into the base service and then calls 4
methods

1. `authorized?` -> handle authorization for the service
2. `execute_on_branch_rule` -> logic for `Projects::BranchRule`
3. `execute_on_all_branches_rule` -> logic for `Projects::AllBranchesRule`
4. `execute_on_all_protected_branches_rule` -> logic for `Projects::AllProtectedBranchesRule`

This should simplify the logic for each child service.