Conform to Anthropic prompt format conditions
We've been seeing errors with message `prompt must start with
"\n\nHuman:" turn after an optional system prompt` because our
conversation truncation logic would sometimes leave an `Assistant:`
message as the beginning of the prompt. To avoid this, we only append
to the prompt conversation blocks that start with a `Human:` message.
We take the chance to follow Anthropic's suggestion of ensuring there
are two line breaks before every `Human:` message.

Ref: https://docs.anthropic.com/claude/reference/prompt-validation