Add `projects[]` to Jira issues GraphQL query
When the param is sent from the URL, we will also include it in the
query sent to the client GraphQL resolvers, which in term send this
to the backend via REST API params.

Similarly to labels, this can accept multiple values, in case we
want to allow filtering by multiple projects at a time.

Also refactored the attributes' order to make them consistent.