Validate syncing of epic work items
This publishes an EpicCreatedEvent and EpicUpdatedEvent which the
ValidateEpicWorkItemSyncWorker subscribes to. As part of the validation
we check if any attribute is different from the work item compared to
the epic.

We then log the attributes that mismatch as a warning. We do explicitly
not log the actual data to prevent leaking any sensitive information.