Remove the trailing slash from the component FQDN
When calculating the component FQDN, we used to add a trailing slash
at the end of the host+port.

When calling a component from a CI config, we check if the call has
a proper FQDN as a prefix. Now, we want to provide the flexibility of
using a variable as the FQDN part of the `include:component`. And we
don't want users to use `${VAR}components` because the `${VAR}` would
have a trailing slash already. So, we are removing it in this MR first.

- Related to https://gitlab.com/gitlab-org/gitlab/-/issues/434260
- Related to https://gitlab.com/gitlab-org/gitlab/-/issues/415638