Implement apply .patch in Keeps::DeleteOldFeatureFlags
The `gitlab-housekeeper` currently has a very simple implementation of
automatically removing feature flags that are past their expected usage
date.

This MR introduces new functionality to `Keeps::DeleteOldFeatureFlags`
where it looks for a `.patch` file matching the same name as the `.yml`
file for the feature flag. If it exists it will apply the patch.

The tool will be used to fully automate the developer workflow for
removing feature flags. Instead of developers needing to remember to
remove their feature flag later they can just add a `.patch` file when
they first create the feature flag. Later on the `gitlab-housekeeper`
will come along and clean it up for them.