Update Transfer Data project blueprint with new Clickhouse fields
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/442907

**Problem**

We discovered that the current approach with recording project_path and
namespace_path in logs has a flaw.

It's possible for clients to update the path of their groups and
projects:
https://docs.gitlab.com/ee/user/group/manage.html#change-a-groups-path.
It means that after update we will lose a connection between old and new
records in Clickhouse.

**Solution**

Use `project_id` and `root_namespace_id` because they are immutable and
won't be affected by the project/group path changes.

Changelog: other