Fix feature flag valid usage examples in errors
The hash used to reference the example in valid_usage! used symbols to
reference values by key. Since we were passing in a string, the lookup
would always fail, and the example in the error message would fail to
display as a result. To fix this, the `type` attribute is cast to a
symbol so that we use the expected type of key (Symbol). The spec has also
been changed to make a strict comparison, so that a regression for this
is caught in the future.