Additional cleanup for external usernames on oAuth and LDAP sign up
Currently if a user joins a GitLab instance via oAuth, LDAP, or SCIM and
their username begins with multiple `--` characters, these characters
are removed to ensure a valid username can be created for the new user.
Illegal characters such as `*&^()` are removed, and illegal extensions
such as `.git` and `.atom` are also removed.

This change extends the behavior to include all leading legal
characters: `-` , `_` and `.` . It also trims from the end of the
potential username, and removes consecutive legal characters so the
resulting username will pass the check for
`Gitlab::Regex.oci_repository_path_regex` .

Changelog: changed