Fix composability of `:published_event` with `:not_publish_event`
Currently it's not possible to use `:published_event` and
`:not_publish_event` together as both matchers are stubbing the
`Gitlab::EventStore#publish` method, so only the first matcher will save
a list of the _published_ events.

To fix that, now we're using `RSpec::PublishedGitlabEventStoreEvents`
to centralized all the events published within an example. This will
enable us to match by multiple events or compose `:published_event` and
`:not_publish_event` in the same spec.

Changelog: fixed