Don't force when generating background migration in housekeeper
This `force: true` option seems to generate the migration even if there
is an existing migration with the exact same name. Additionally it
deletes the existing migration with the same name. This is probably not
something we ever intended to do so I'm removing it.

At some point we may want to update this to skip finalizing a BBM if a
migration with a matching name exists (or generate a guaranteed unique
name) but it's not clear to me if we actually have a use case for this
now. Also there are plans to be more strict about not re-using batched
background migrations so possibly the matching name won't be very common
in future.