Use primary DB for computing updates
The DB load balancer _switches_ to
the primary when it detects a transaction or
other statements that modify data or schema.

This behavior might lead to the issuable update services
to read lagging data from a replica to compute the desired
updated states.

While it's desirable to use a transaction in conjunction with locking
to compute updates, the issuable services have not been designed
to retroactively accomodate such change.

To ensure that the issuable services compute new states
using the latest data, the update method is updated to
switch to read from the primary from the beginning.
This won't address the existing data race conditions
but might abate them.