Add CI_COMPONENT_FQDN variable and usage instructions
Currently, the FQDN is written hard-coded into components and
pipeline configs. This will cause more changes needed when self-hosted
instances mirror a component from gitlab.com and try to use it.

As a result, we are introducing the CI_COMPONENT_FQDN variable.
It will be the suggested way when creating a component and use it.

Changelog: added