Add `initial` &  `latest` pipeline IDs to `vulnerability_occurrences`
We want to drop the `vulnerability_occurrence_pipelines` table. In
order to do this, we still need the first and latest pipeline ID
stored somewhere to support the existing feature-set

Follow-up Work
---------------------------

We will need to add loose foreign keys and model association. Both of
these are blocked on index creation, which has to be done via the
[async process][0]

That work is tracked in the following tasks:

- https://gitlab.com/gitlab-org/gitlab/-/work_items/443282
- https://gitlab.com/gitlab-org/gitlab/-/work_items/443283
- https://gitlab.com/gitlab-org/gitlab/-/work_items/443284

---

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/422382
Resolves https://gitlab.com/gitlab-org/gitlab/-/work_items/443281
Resolves https://gitlab.com/gitlab-org/gitlab/-/work_items/443280

[0]:https://docs.gitlab.com/ee/development/database/adding_database_indexes.html#create-indexes-asynchronously

Changelog: added