Add a worker to unstick canceling jobs
Adds a worker that will move jobs with a canceling status to canceled
if they have been stuck in canceling for too long. This is what we do
with other non-complete statuses where jobs have been stuck in that status.
Since we rely on an external components to message back a failure we want to
ensure that even if that dosn't happen, jobs will eventually arrive at
a complete status.