Handle the unadvertised object error
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/427699

**Problem**

When user imports a repository from the BitBucket Server, we are trying
to fetch missing commits. See more information:
https://gitlab.com/gitlab-org/gitlab/-/issues/356240.

But we discovered a problem with repositories that have submodules. The
commit fetching process fails for them, because we have to fetch
submodules first.

**Solution**

As a workaround to unblock the import process, we can skip loading of
specific commits when the repository has submodules.