Include approval_rule_id in uniqueness scope
Users should be able to approve a deployment once for each rule they can
satisfy. This covers the use case of approving for multiple groups.
There will be a follow up to to drop the existing unique index, and
replace it with an index that includes approval_rule_id Follow up issue:
https://gitlab.com/gitlab-org/gitlab/-/issues/442601