Remove legacy finding modal and vulnerability_details
This MR removes the legacy finding modal completely, together with the
only remaining child component within in, the vulnerability_details.vue.
In addition, the specs vulnerability_details_spec, the spec snapshot,
a helper get_file_location, translations, and a lot of unused mock_data
are removed. Lastly, an unused part of custom CSS is removed.