Update pipeline_triggers plan limits to 25k for trials and opensource
Currently the pipeline_triggers limits across plans
on GitLab.com are not consistent. This is because either new
plans were introduced or the limits were not migrated when
the plans were renamed. This MR introduces a migration to
update the values on GitLab.com to 25k for non-paid plans.

Changelog: fixed