Log diffs_files_count when MR diffs and commit are accessed
To be able to give an estimate on how much storage we are going
to use when we cache diffs HTML, we need to get some data on
how many files are there in diffs.

We add this logging metadata to `MergeRequestsController#diffs` and
`CommitController#show` actions since those are the actions that
we want to track.