Fix the position of the version header in changelog
Contibutes to https://gitlab.com/gitlab-org/gitlab/-/issues/437616

**Problem**

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/144738 allowed
users to provide version with `v` at the beginning. However, the
changelog generator was not adapted to accept it.

As a result, it couldn't correctly position a new version header in
CHANGELOG.md file.

**Solution**

Remove the prefix `v` from the version name.