Document Container Scanning integration with CVS
This documents the component types supported by Continuous Vulnerability
Scanning, and mentions Container Scanning as a source of these components.
It also documents Trivy DB updates as a trigger for the scans performed,
and links to the issue tracking support for the remaining `apk` and
`rpm` component types not supported in the MVC release Continuous
Vulnerability Scanning for Container Scanning.