Extend epic work item validation checks
As we now sync more attributes to the work_item, this adds more checks
to validate that the epic and the work item are in sync.

We now check for the hierarchy, the start_date, the due_date, and
the related epic links.

This chnage also introduces the optional `strict_equal` parameter that
is `false` by default now. When `strict_equal` is `true` we expect that
all relationships are synced. This is the case when we have backfilled
all data.