Allow Vite HMR server to work in feature specs
In browser feature tests with, we start Chrome with a proxy server to
reject any external network access. However, this also blocked the
Vite HMR connection, which needs to connect to the GDK hostname via a
Websocket, as configured in the `server.hmr.host` Vite setting. In
tests, `Gitlab.config.gitlab.host` is usually set to `localhost`
instead of the GDK server hostname, so Chrome was always rejecting
Vite HMR access. We fix this by storing the HMR hostname in
`ViteServer.env` and adding this host in the proxy bypass list.

This appears to fix the Content Security Policy blocking issues with
running feature tests. If the HMR is not allowed, the Vite client will
attempt to ping the HMR server with `https://` instead of `wss://`, as
described in
https://github.com/vitejs/vite/pull/16056#issuecomment-1973640700.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/443737