Prevent adding new members when block seat overages is enabled
When block seat overages is enabled for a root group, total billable
members cannot exceed the number of seats in the subscription.
Reject requests to add new members to groups or projects if block
seat overages is enabled for the root group and all the seats in the
subscription are full.
This change is behind a feature flag.