Validate that sensitive emails are sent to a single recipient
This is a defense-in-depth measure. It introduces a "time of use"
check to validate that, when appropriate, an email is only being sent
to a single email address. This should already be the case but, in the
event a bug is introduced further up the stack in the future, this
should prevent potential security vulnerabilities<sup>1</sup>.

This check is performed for:

- All Devise emails<sup>2</sup>:
    - `confirmation_instructions`
    - `reset_password_instructions`
    - `unlock_instructions`
    - `email_changed`
    - `password_change`
- GitLab's extra emails in `DeviseMailer`:
    - `password_change_by_admin`
    - `user_admin_approval`
    - `reset_password_instructions`
    - `email_changed`
- Identity Verification<sup>3</sup> emails:
    - `verification_instructions_email`
    - `confirmation_instructions_email`

None of these emails should be sent to more than one person, or have
other recipients in the CC or BCC fields. An exception will be raised
if this occurs.

As many of these emails are sent in a delayed manner, the exceptions
will be handled by ActiveJob.

References:

1. https://gitlab.com/gitlab-org/gitlab/-/issues/436084
2. GitLab's `DeviseMailer` is set as Devise's mailer in
`config/initializers/8_devise.rb:22`
3. https://docs.gitlab.com/ee/security/identity_verification.html
4. https://gitlab.com/gitlab-org/gitlab/-/issues/442831