Allow adding members already in group when using block seat overages
If block seat overages is enabled, and all the seats in the group are
full, then accept a request to add a member if the member is
already in the group.
Prior to this change, the request would be rejected with an error
message that all the seats are full.
Block seat overages is behind a feature flag.