Reload HEAD diff when MR is marked as mergeable
When syncing code owner rules, we are only looking for HEAD diff
if MR is mergeable.

However, before this fix, we reload the HEAD diff and sync code
owner rules before a MR gets marked as mergeable. That results to
code owner rules not being based on the HEAD diff even if the
MR is mergeable.

To fix it, we reload the HEAD diff by calling
`MergeRequests::ReloadMergeHeadDiffService` after a MR gets marked
as mergeable.

This way, if MR is mergeable, the HEAD diff is reloaded and code
owner rules gets synced based off the updated HEAD diff. If MR
is not mergeable, we don't reload the HEAD diff since we don't
display it if MR is not mergeable.

Changelog: fixed