Cell 1.0 initial unsupported features
What
---
Create a list of the features that will not be supported in Cell 1.0
initial deployment.

Why
---
We don't have a single source of truth of what is going to be supported
and what is not. This doesn't mean that we will not support them in
Cells 1.0, but will add them after we start onboarding some new
customers.

Ideally, this should serve as a single source of truth so that
engineering, product, and sales can use this list.

Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>