Correct loading state after failing prompt
We currently only update the loading state of duo chat after the an
assistant message has been sent from the backend.

In the case of banned AI, we actually never get an assistant message,
but an error directly in the prompt leading to a indefinite loading
state.

Now we double check that if the user message has `errors`, we just set
loading to false.