Synchronously add indexes for new `vulnerability_occurrences` cols
We want to drop the `vulnerability_occurrence_pipelines` table. In
order to do this, we still need the first and latest pipeline ID
stored somewhere to support the existing feature-set

The columns and asynchronous index creation were added in a
[previous MR][0]

This change adds the synchronous index migration to complete the index
creation [process][1]

Follow-up Work
---------------------------

We will need to add loose foreign keys and model association. Both of
these are blocked on this index creation

That work is tracked in the following tasks:

- https://gitlab.com/gitlab-org/gitlab/-/work_items/443283
- https://gitlab.com/gitlab-org/gitlab/-/work_items/443284

---

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/422382
Resolves https://gitlab.com/gitlab-org/gitlab/-/work_items/443282

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/144697
[1]:https://docs.gitlab.com/ee/development/database/adding_database_indexes.html#create-indexes-asynchronously

Changelog: added