Make development seed for application settings idempotent
This enables users to run `FILTER=application_settings bin/rake
db:seed_fu` as many times as they want, and allows retries to happen
in CI if needed.

This also reduces the number of updates to the application settings.
Each `update!` causes a validation to occur. Each URL will be
validated by resolving its DNS address, which could be slow or time
out.

Relates to
https://gitlab.com/gitlab-org/gitlab-development-kit/-/issues/2062