Fix moving an issue to a synced epic
In the `EpicIssue` validations we check if there is a correlating
`WorkItems::ParentLink`. If there is one, we make sure that the
`ParentLink` reflects the hierarchy we have set via the `EpicIssue`
hierarchy. This is a measurement to not get the legacy epics and the new
work item epics out of sync.

However, the validation hinders us to move an issue from one epic to
another, when a synced work item epic and `WorkItems::ParentLink`
already exists.

As a fix, we skip the validation to change the `EpicLink` and
`WorkItems::ParentLink` within a transaction.