Improve error message when job is missing an artifact
In https://gitlab.com/gitlab-org/gitlab/-/pipelines/1205422081,
we saw the error "Unable to create pipeline" with the message
"Job does not have artifact". The pipeline depends on 4
jobs, and it's not clear which one is the problem. Improve
this by adding the job name and ID to the error messages.

Changelog: changed