Fix sidebar padding problem for issue lists
Issue lists never have a collapsed sidebar, but they had previously
shown a gap at the right to accommodate a collapsed sidebar at small
viewport sizes. This fixes that, and fixes a subtle bug that caused that
right side padding to be re-added when the issue bulk edit sidebar was
opened.

Changelog: fixed