Fix failing spec/scripts/internal_events/cli_spec.rb test
`spec/scripts/internal_events/cli_spec.rb` attempts to look at a
mocked TTY output to see if the CLI worked as behaved. However, the
test started to fail because the output grew to 622 lines. However,
the test looked only at the last 320 lines, but not all the expected
outputs were contained in that window.

To fix this test, just look through the entire buffer instead
of trying to limit by hundreds of rows.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/448636