Mask membership source if the current user cannot access the source
As part of https://gitlab.com/gitlab-org/gitlab/-/issues/219230, we will
start showing the invited group members on the project/group members
page. However, the current user might not have access to the invited
group so depending on different cases we are hiding the source of
the invited group member or the invited member itself.

If the invited group is public then we will always show its members on
the shared project/group page. But if it's private and it's invited to a
public group/project then the following cases are there:

1. Current user is unauthenticated - The user won't see the members from
the invited group on the shared project/group members page.
2. Current user is a non-member of the invited group and the shared
group/project - This is the same as point 1.
3. Current user is a member of the shared group/project but not of the
invited group - The user will see the members of the invited group
but the source of membership will be masked.
4. Current user is a member of the invited group - The user will be able
to see the source of membership.
5. Current user is the maintainer/owner of the shared project or owner
of the group - The user can see the source of membership to manage
the project/group memberships.