Change "projects[]" param to "project"
Change in implementation since we will only allow filtering by
a single project at a time. The `projects[]` syntax with multiple
values also wasn't consistent with labels as `labels[]` implied an AND query while projects (if supported) would always be an OR query.

This change is behind a feature flag `jira_multiple_project_keys`.