Properly handle email invites for block seat overages
When block seat overages is enabled for a group, then the number of
billable users cannot exceed the number of seats in the subscription.
Prevent email invites from being added to the namespace in excess of
the total number of seats in the subscription.
This feature is behind a feature flag.