Commits view: Add "$file did not exist on $ref" error
When a given path is not present on a valid reference, show an error
message and redirect to the commits view instead of redirecting to a 404
page.

That makes the commits view match the behavior of the tree view.

Updates `spec/controllers/projects/commits_controller_spec.rb` with the
new logic.