Refactor specs to not use vulnerabilities_finding_pipeline directly
We are working on an [epic][0] to drop the
`vulnerability_occurrence_pipelines` table.

When we eventually migrate off of this table/model, there are many
specs that will need to be updated to no longer use the
`:vulnerabilities_finding_pipeline` factory

---

This refactor to the `Vulnerability::Finding` factory centralizes the
pipeline association to the `FactoryBot` layer, rather than having the
association happen at the spec level.

This change will prevent the migration MR from including a lot of
tangentially related spec changes

[0]:https://gitlab.com/groups/gitlab-org/-/epics/11241

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/422384
resolves: https://gitlab.com/gitlab-org/gitlab/-/work_items/450524