Upgrade gitlab-fog-azure-rm to v1.9.1
`send_file_upload` attempts to set `response-content-type` via a
`query` parameter, but this previously only worked for fog-aws and
fog-google. For Azure, we need to convert this `query` parameter for
the Azure Blob Storage client, which uses the `content_type` argument
to a SAS `rsct` parameter.  This is done in
https://gitlab.com/gitlab-org/ruby/gems/gitlab-fog-azure-rm/-/merge_requests/44.

Previously the dependency proxy did not work since the
`gitlab-fog-azure-rm` did not convert this `query` argument properly.

This change makes the dependency proxy work with Azure Blob Storage.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/352551

Changelog: fixed