Add required token role level
Creating a project token defaults to the guest role. Creating a token with the api scope and guest role is insufficient as it does not allow checking out the code of the repository, the bootstrap git checkout step fails with a permission error.  A project token with the api scope and developer role works.