Add usage_count columns to catalog tables
Adds 'last_30_day_usage_count' column to
'catalog_resources' and 'catalog_resource_components'.
The column is to be later populated by a daily worker
that aggregates the last 30 days of data. The usage
count is currently defined as the number of unique
projects that included the respective component in a
pipeline.

Changelog: added