Log if pipeline creation limit might be exceeded
Previously if a user pushed more than 4 refs to a pipeline, the
PostReceive Sidekiq job would pick the first 4 and ignore the
rest. This commit adds a log message in `application_json` to warn the
admin this might be happening. Note that we can't be sure if a given
ref would actually have a pipeline created for it since the CI rules
would need to be evaluated to determine that.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/350886

Changelog: changed