Fix vuln findings pagination tests
These tests are checking for the presence of the `X-Total` pagination
header but we do not actually want these on this endoint.
The endpoint uses pagination without counts which does not produce this
header.

Other changes on this MR have removed the headers and highlighted these
outdated tests, so we're removing them here.