Add `spec` column to `catalog_resource_components`
When we add CI steps to the CI Catalog, we'll need to store the entire
`spec` header section in the table. This commit is the first step
towards migrating from the `inputs` field to a new `spec` field that
will support the needs of both steps and components.

Changelog: added
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/443662