Save component metadata in `spec` column
We've been saving CI components inputs metadata in a column named
`inputs`, but when we introduce CI steps we'll also need to save
`outputs`. We've introduced a `spec` column that will contain the entire
`spec` value for all CI component types. This commit updates the code to
save to the `spec` column instead of `inputs`. It also fetches first
from `spec`, and then from `inputs` if `spec` is `nil`.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/443662