Ensure rubocop files end up in GDK QA image
This is for
https://gitlab.com/gitlab-org/database-team/gitlab-com-database-testing/-/merge_requests/188
where we are trying to use this GDK image for running housekeeper.
Housekeeper needs a full gitlab rails environment with some database and
the ability to run rubocop.

It's possible that there may be a better docker image that we could be
using but this one seemed close enough to match our needs if it just had
rubocop.